+++ emperor/support_files/js/animate.js
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 define([
     'underscore',
     'trajectory'
@@ -7,6 +8,90 @@ function(_, trajectory) {
     trajectory.getSampleNamesAndDataForSortedTrajectories;
   var getMinimumDelta = trajectory.getMinimumDelta;
   var TrajectoryOfSamples = trajectory.TrajectoryOfSamples;
+=======
+/**
+ *
+ * @author Yoshiki Vazquez Baeza
+ * @copyright Copyright 2013, The Emperor Project
+ * @credits Yoshiki Vazquez Baeza
+ * @license BSD
+ * @version 0.9.61
+ * @maintainer Yoshiki Vazquez Baeza
+ * @email yoshiki89@gmail.com
+ * @status Development
+ *
+ */
+
+/**
+ *
+ * @name AnimationDirector
+ *
+ * @class This class represents an animation director for PCoA plots in the
+ * Emperor visualization software.
+ *
+ * @property {float} [minimumDelta=null] A floating point value determining
+ * what the minimum separation between samples along the gradients is. Will be
+ * null until it is initialized to the values according to the inputed data.
+ * @property {int} [maximumTrajectoryLength=null] Maximum length that the
+ * groups
+ * of samples have along a gradient.
+ * @property {int} [currentFrame=-1] The current frame being served by the
+ * director.
+ * @property {Array} [trajectories=Array(0)] Array where each element in the
+ * trajectory is a trajectory with the interpolated points in it.
+ *
+ */
+
+/**
+ *
+ * @name AnimationDirector
+ *
+ * @class This object represents an animation director, as the name implies, is
+ * an object that manages an animation. Takes the for a plot (mapping file and
+ * coordinates) as well as the metadata categories we want to animate over.
+ * This object gets called in the main emperor module when an animation starts
+ * and an instance will only be alive for one animation cycle i. e. until the
+ * cycle hits the final frame of the animation.
+ *
+ * @param {mappingFileHeaders} an Array of strings containing metadata mapping
+ * file headers (required).
+ * @param {mappingFileData} an Array where the indices are sample identifiers
+ * and each of the contained elements is an Array of strings where the first
+ * element corresponds to the first data for the first column in the mapping
+ * file (mappingFileHeaders) (required).
+ * @param {coordinatesData} an Array of Objects where the indices are the
+ * sample identifiers and each of the objects has the following properties: x,
+ * y, z, name, color, P1, P2, P3, ... PN where N is the number of dimensions in
+ * this dataset (required).
+ * @param {gradientCategory} a string with the name of the mapping file header
+ * where the data that spreads the samples over a gradient is contained,
+ * usually time or days_since_epoch. Note that this should be an all numeric
+ * category (required).
+ * @param {trajectoryCategory} a string with the name of the mapping file
+ * header where the data that groups the samples is contained, this will
+ * usually be BODY_SITE, HOST_SUBJECT_ID, etc. (required).
+ * @param {speed} Positive real number determining the speed of an animation,
+ * this is reflected in the number of frames produced for each time interval.
+ *
+ * @return returns an animation director if the parameters passed in were all
+ * valid.
+ *
+ * Note that this class will raise an Error in any of the following cases:
+ * - One of the input arguments is undefined.
+ * - If gradientCategory is not in the mappingFileHeaders.
+ * - If trajectoryCategory is not in the mappingFileHeaders.
+ *
+ */
+function AnimationDirector(mappingFileHeaders, mappingFileData, coordinatesData,
+                           gradientCategory, trajectoryCategory, speed){
+
+  // all arguments are required
+  if (mappingFileHeaders === undefined || mappingFileData === undefined ||
+    coordinatesData === undefined || gradientCategory === undefined ||
+    trajectoryCategory === undefined) {
+    throw new Error("All arguments are required");
+  }
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
 
   /**
    *
@@ -73,6 +158,7 @@ function(_, trajectory) {
                       ' file');
     }
 
+<<<<<<< HEAD
     // guard against logical problems with the trajectory object
     if (speed <= 0) {
       throw new Error('The animation speed cannot be less than or equal to ' +
@@ -154,6 +240,54 @@ function(_, trajectory) {
     this.getMaximumTrajectoryLength();
 
     return this;
+=======
+  // guard against logical problems with the trajectory object
+  if (speed <= 0) {
+    throw new Error("The animation speed cannot be less than or equal to zero");
+  }
+
+  this.mappingFileHeaders = mappingFileHeaders;
+  this.mappingFileData = mappingFileData;
+  this.coordinatesData = coordinatesData;
+  this.gradientCategory = gradientCategory;
+  this.trajectoryCategory = trajectoryCategory;
+
+  this.minimumDelta = null;
+  this.maximumTrajectoryLength = null;
+  this.currentFrame = -1;
+  this.trajectories = new Array();
+  this.speed = speed;
+
+  this.initializeTrajectories();
+  this.getMaximumTrajectoryLength();
+
+  return this;
+}
+
+/**
+ *
+ * Initializes the trajectories that the director manages.
+ *
+ */
+AnimationDirector.prototype.initializeTrajectories = function(){
+
+  var chewedData = null, trajectoryBuffer = null, minimumDelta;
+  var sampleNamesBuffer = new Array(), gradientPointsBuffer = new Array();
+  var coordinatesBuffer = new Array();
+  var chewedDataBuffer = null;
+
+  // frames we want projected in the trajectory's interval
+  var n = Math.floor((1 / (this.speed)) * 10);
+
+  // compute a dictionary from where we will extract the germane data
+  chewedData = getSampleNamesAndDataForSortedTrajectories(
+      this.mappingFileHeaders, this.mappingFileData, this.coordinatesData,
+      this.trajectoryCategory, this.gradientCategory);
+
+  if (chewedData === null){
+    throw new Error("Error initializing the trajectories, could not "+
+                    "compute the data");
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
   }
 
   /**
@@ -246,8 +380,14 @@ function(_, trajectory) {
       this._computeN();
     }
 
+<<<<<<< HEAD
     return this.maximumTrajectoryLength;
   };
+=======
+    // create the trajectory object
+    trajectoryBuffer = new TrajectoryOfSamples(sampleNamesBuffer, key,
+        gradientPointsBuffer, coordinatesBuffer, this.minimumDelta, n);
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
 
   /**
    *