+++ tests/all_tests.py
@@ -25,6 +25,70 @@ from os.path import join, abspath, dirname, split
 
 from emperor import __version__
 
+<<<<<<< HEAD
+=======
+from qcli.util import qcli_system_call
+from qcli.test import run_script_usage_tests
+from qcli.option_parsing import parse_command_line_parameters, make_option
+
+script_info = {}
+script_info['brief_description'] = ""
+script_info['script_description'] = ""
+script_info['script_usage'] = [("", "", "")]
+script_info['output_description'] = ""
+script_info['required_options'] = []
+script_info['optional_options'] = [
+    make_option('--suppress_unit_tests', action='store_true', help='suppress '
+                " execution of Emperor's unit tests [default: %default]",
+                default=False),
+    make_option('--suppress_script_usage_tests', action='store_true',
+                help="suppress Emperor's script usage tests [default: "
+                "%default]", default=False),
+    make_option('--suppress_javascript_unit_tests', action='store_true',
+                help='suppress Emperor\'s JavaScript unit tests [default: '
+                '%default]', default=False),
+    make_option('--unittest_glob', help='wildcard pattern to match the unit '
+                'tests to execute [default: %default]', default=None),
+    make_option('--script_usage_tests', help='comma-separated list of the '
+                'script usage tests to execute [default: run all]',
+                default=None),
+    make_option('-p', '--temp_filepath', type='existing_path', help='temporary'
+                ' directory where the script usage tests will be executed',
+                default='/tmp/'),
+    make_option('--emperor_scripts_dir', help='filepath where the scripts are'
+                ' stored', type='existing_path', default=None)
+]
+script_info['version'] = __version__
+script_info['help_on_no_arguments'] = False
+
+
+def main():
+    option_parser, opts, args = parse_command_line_parameters(**script_info)
+
+    unittest_glob = opts.unittest_glob
+    temp_filepath = opts.temp_filepath
+    script_usage_tests = opts.script_usage_tests
+    suppress_unit_tests = opts.suppress_unit_tests
+    suppress_script_usage_tests = opts.suppress_script_usage_tests
+    suppress_javascript_unit_tests = opts.suppress_javascript_unit_tests
+
+    # since the test data is in the tests folder just add scripts_test_data
+    emperor_test_data_dir = join(abspath(dirname(__file__)),
+                                 'scripts_test_data/')
+
+    # offer the option for the user to pass the scripts dir from the command
+    # line since there is no other way to get the scripts dir. If not provided
+    # the base structure of the repository will be assumed. Note that for both
+    # cases we are using absolute paths, to avoid unwanted failures.
+    if opts.emperor_scripts_dir is None:
+        emperor_scripts_dir = abspath(join(get_emperor_project_dir(),
+                                           'scripts/'))
+
+        # let's try to guess cases for qiime-deploy type of installs
+        if get_emperor_project_dir().endswith('/lib'):
+            emperor_scripts_dir = abspath(join(get_emperor_project_dir()[:-3],
+                                               'scripts/'))
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
 
 def console(cmd, stdout=None, stderr=None):
     """Small subprocess helper function
@@ -61,9 +125,15 @@ def test(suppress_unit_tests, suppress_javascript_unit_tests, unittest_glob):
     """
 
     # make a sanity check
+<<<<<<< HEAD
     if suppress_unit_tests and suppress_javascript_unit_tests:
         raise click.UsageError("All tests have been suppresed. Nothing to "
                                "run.")
+=======
+    if (suppress_unit_tests and suppress_script_usage_tests and
+            suppress_javascript_unit_tests):
+        option_parser.error("All tests have been suppresed. Nothing to run.")
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
 
     test_dir = abspath(dirname(__file__))
 
@@ -101,6 +171,37 @@ def test(suppress_unit_tests, suppress_javascript_unit_tests, unittest_glob):
                 else:
                     bad_tests.append(unittest_name)
 
+<<<<<<< HEAD
+=======
+    script_usage_failures = 0
+
+    # choose to run some of the script usage tests or all the available ones
+    if (not suppress_script_usage_tests and exists(emperor_test_data_dir) and
+            exists(emperor_scripts_dir)):
+        if script_usage_tests is not None:
+            script_tests = script_usage_tests.split(',')
+        else:
+            script_tests = None
+
+        initial_working_directory = getcwd()
+
+        # Run the script usage testing functionality; note that depending on
+        # the module where this was imported, the name of the arguments will
+        # change that's the reason why I added the name of the arguments in
+        # here
+        script_usage_result_summary, script_usage_failures = \
+            run_script_usage_tests(emperor_test_data_dir,  # test_data_dir
+                                   emperor_scripts_dir,    # scripts_dir
+                                   temp_filepath,          # working_dir
+                                   True,                   # verbose
+                                   script_tests,           # tests
+                                   None,                   # failure_log_fp
+                                   False)                  # force_overwrite
+
+        # running script usage tests breaks the current working directory
+        chdir(initial_working_directory)
+
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
     if not suppress_javascript_unit_tests:
         print("JavaScript Test Suite")
         runner = join(test_dir, 'javascript_tests', 'runner.js')
@@ -122,17 +223,40 @@ def test(suppress_unit_tests, suppress_javascript_unit_tests, unittest_glob):
     if not suppress_unit_tests:
         print("\nUnit test result summary\n------------------------\n")
         if bad_tests:
+<<<<<<< HEAD
             print("\nThe following unit tests failed.\n%s"
                   % '\n'.join(bad_tests))
 
         if missing_application_tests:
             print("\nThe following unit tests failed, in part or whole due "
+=======
+            print("\nFailed the following unit tests.\n%s"
+                  % '\n'.join(bad_tests))
+
+        if missing_application_tests:
+            print("\nFailed the following unit tests, in part or whole due "
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
                   "to missing external applications.\nDepending on the "
                   "Emperor features you plan to use, this may not be "
                   "critical.\n%s" % '\n'.join(missing_application_tests))
 
         if not(missing_application_tests or bad_tests):
+<<<<<<< HEAD
             print("\nAll unit tests passed.\n")
+=======
+            print "\nAll unit tests passed.\n"
+
+    if not suppress_script_usage_tests:
+        if exists(emperor_test_data_dir) and exists(emperor_scripts_dir):
+            print("\nScript usage test result summary"
+                  "\n--------------------------------\n")
+            print script_usage_result_summary
+        else:
+            print("\nCould not run script usage tests.\nThe Emperor scripts "
+                  "directory could not be automatically located, try "
+                  "supplying it manually using the --emperor_scripts_dir "
+                  "option.")
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
 
     if not suppress_javascript_unit_tests:
         print('\nJavaScript unit tests result summary\n'
@@ -145,7 +269,11 @@ def test(suppress_unit_tests, suppress_javascript_unit_tests, unittest_glob):
     # In case there were no failures of any type, exit with a return code of 0
     return_code = 1
     if (len(bad_tests) == 0 and len(missing_application_tests) == 0 and
+<<<<<<< HEAD
        javascript_tests_passed):
+=======
+            script_usage_failures == 0 and javascript_tests_passed):
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
         return_code = 0
 
     exit(return_code)