+++ ChangeLog.md
@@ -1,6 +1,7 @@
 Emperor ChangeLog
 =================
 
+<<<<<<< HEAD
 # Emperor 1.0
 
 Introducing Emperor 1.0, an improved and new version of Emperor. Including
@@ -8,6 +9,15 @@ a stable Python API and JavaScript API.
 
 Emperor 0.9.51-dev
 ------------------
+=======
+Emperor 0.9.61 (11 Apr 2018)
+----------------------------
+
+* Maintenance release ([#631](https://github.com/biocore/emperor/issues/631)).
+
+Emperor 0.9.60 (2 Nov 2016)
+---------------------------
+>>>>>>> 32f0c53e72a9543ffd3a6edb1b8772d5ca73def1
 
 ### Bug Fixes
 * Fix problem where the taxonomic vectors weren't scaled properly when the scale coordinates button was toggled  ([#386](https://github.com/biocore/emperor/issues/373).
@@ -15,6 +25,8 @@ Emperor 0.9.51-dev
 * Fix bug where Emperor would try to animate trajectories with a single timepoint i.e. a single unique value in the gradient category.
 * Fix bug where the output saved to `--biplot_fp` would list the principal coordinate axes starting at zero instead of one ([#389](https://github.com/biocore/emperor/issues/389)).
 * Remove misleading "QIIME version" from `index.html` ([#447](https://github.com/biocore/emperor/issues/447)).
+* Fix problem where the speed slider in the *Animations* tab would not work correctly ([#546](https://github.com/biocore/emperor/pull/546)).
+* Fix problem bug that would prevent users from using the *Visibility* and *Scaling* controllers ([#420](https://github.com/biocore/emperor/issues/420)).
 
 ### Miscellaneous
 
@@ -29,6 +41,7 @@ Emperor 0.9.51-dev
 
 * Add a slider bar under the `Labels` tab to select different taxonomies for display.
 * Biplots now include arrows, which represent taxa variances. These can be turned off using the `Taxa arrow visibility` checkbox.
+* Add new tutorial on how to create animations ([#547](https://github.com/biocore/emperor/pull/547)).
 
 Emperor 0.9.51 (5 Feb 2015)
 ---------------------------